refactor: getAttackStatus
